no float

+++ src/core/lib/math/transfrm.h
@@ -84,7 +84,7 @@ namespace lbcrypto {
 		/*Ring dimension factor calculates the ratio between the cyclotomic order of the root of unity table
 			  that was generated originally and the cyclotomic order of the current VecType. The twiddle table
 			  for lower cyclotomic orders is smaller. This trick only works for powers of two cyclotomics.*/
-		float ringDimensionFactor = ((float)rootOfUnityTable.GetLength()) / (float)cycloOrder;
+		int ringDimensionFactor = rootOfUnityTable.GetLength() / cycloOrder;
 		DEBUG("table size " << rootOfUnityTable.GetLength());
 		DEBUG("ring dimension factor " << ringDimensionFactor);
